# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VigiaApp API
  version: 1.1.0
  description: |
    API REST de VigiaApp. Parte de la especificación v1.0 (`VigiaApp/docs/especificacion_api_db.md`);
    las diferencias con esa versión están en `docs/CAMBIOS_CLIENTE.md`.

    - Todas las respuestas son `application/json; charset=utf-8` (RNF-04).
    - Claves en `snake_case`; los booleanos viajan como `true`/`false` (RNF-07).
    - Fechas con hora: ISO-8601 en UTC con `Z` y sin milisegundos (`2026-09-23T18:35:00Z`).
      Fechas de calendario: `AAAA-MM-DD`. Horas: `HH:MM`.
    - Todas las rutas, salvo `/auth/login`, `/auth/clave` y `/auth/recuperar`, exigen `Authorization: Bearer <token>`.
    - Cualquier respuesta que no sea 2xx trae el cuerpo `Error`.
servers:
  - url: https://{dominio}/api/v1
    variables:
      dominio:
        default: api.vigiaapp.example
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
security:
  - bearer: []
...
paths:
...
  /auth/login:
    post:
      tags: [Sesión]
      summary: Iniciar sesión con email y contraseña
      description: Reemplaza a `/auth/codigo` + `/auth/verificar` de la spec v1.0. El vecino debe estar activo y haber elegido contraseña.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: maria@vigia.test }
                password: { type: string, example: vigia1234 }
      responses:
        '200': { $ref: '#/components/responses/Sesion' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '429': { $ref: '#/components/responses/DemasiadosIntentos' }
  /auth/clave:
    post:
      tags: [Sesión]
      summary: Elegir contraseña con el enlace del correo (invitación o recuperación)
      description: |
        El correo trae `APP_URL_CLAVE?token=<64 hex>`. La app toma el `token`, pide la contraseña y llama acá.
        El enlace es de un solo uso (invitación: 7 días; recuperación: 1 hora). Una recuperación además cierra
        todas las sesiones previas del vecino. Responde con una sesión nueva, igual que `/auth/login`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string, pattern: '^[0-9a-f]{64}$' }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '200': { $ref: '#/components/responses/Sesion' }
        '400':
          description: Contraseña fuera de rango, o enlace inválido, vencido o ya usado.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429': { $ref: '#/components/responses/DemasiadosIntentos' }
  /auth/recuperar:
    post:
      tags: [Sesión]
      summary: Pedir el correo para restablecer la contraseña
      description: Responde 204 exista o no el email, para no revelar quién está registrado.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '204': { description: 'Si el email corresponde a un vecino activo, se le envió el enlace.' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '429': { $ref: '#/components/responses/DemasiadosIntentos' }
  /auth/logout:
    post:
      tags: [Sesión]
      summary: Cerrar la sesión actual
      responses:
        '204': { description: El token dejó de valer. }
        '401': { $ref: '#/components/responses/NoAutorizado' }

  # ---------------------------------------------------------------- Perfil
...
  /alertas:
    get:
      tags: [Alertas]
      summary: Alertas del barrio (por defecto, las activas)
      description: |
        Sin push, la app se entera de una alerta nueva haciendo polling de esta ruta (sugerido: cada 10 s
        mientras está abierta). Últimas 50.
      parameters:
        - { name: estado, in: query, required: false, schema: { $ref: '#/components/schemas/EstadoAlerta/properties/estado', default: activa } }
      responses:
        '200':
          description: Alertas
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/EstadoAlerta' } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
    post:
      tags: [Alertas]
      summary: Disparar el SOS (RF-01)
      description: |
        Crea la alerta y una fila `notificado` por cada otro vecino activo, en una sola transacción (RNF-01).
        Si el vecino ya tiene una alerta activa, devuelve esa con 200 y no vuelve a avisar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lat: { type: [number, 'null'], minimum: -90, maximum: 90 }
                lng: { type: [number, 'null'], minimum: -180, maximum: 180 }
      responses:
        '201':
          description: Alerta creada
          content: { application/json: { schema: { $ref: '#/components/schemas/AlertaCerrada' }, example: { id: 17, estado: activa } } }
        '200':
          description: Ya tenía una alerta activa; es esa.
          content: { application/json: { schema: { $ref: '#/components/schemas/AlertaCerrada' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /alertas/{id}:
    get:
      tags: [Alertas]
      summary: Triple confirmación (RF-02). La app la consulta cada 2 s.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Estado de la alerta
          content: { application/json: { schema: { $ref: '#/components/schemas/EstadoAlerta' } } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /alertas/{id}/cancelar:
    post:
      tags: [Alertas]
      summary: Cancelar una falsa alarma (RF-03)
      description: Solo quien la disparó (si no, 404) y solo si sigue activa (si no, 409).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Cancelada
          content: { application/json: { schema: { $ref: '#/components/schemas/AlertaCerrada' }, example: { id: 17, estado: cancelada } } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /alertas/{id}/resolver:
    post:
      tags: [Alertas]
      summary: Dar la emergencia por terminada
      description: Quien la disparó o el comité. Solo si sigue activa.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Resuelta
          content: { application/json: { schema: { $ref: '#/components/schemas/AlertaCerrada' }, example: { id: 17, estado: resuelta } } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /alertas/{id}/en-camino:
    post:
      tags: [Alertas]
      summary: '"Voy en camino" (o retractarse)'
      description: Solo los vecinos avisados (quien disparó la alerta no lo es → 404). La alerta tiene que estar activa (409).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [en_camino]
              properties:
                en_camino: { type: boolean }
      responses:
        '200':
          description: Estado de la alerta actualizado
          content: { application/json: { schema: { $ref: '#/components/schemas/EstadoAlerta' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /alertas/{id}/confirmar-911:
    post:
      tags: [Alertas, Comité]
      summary: Confirmar que la central 911 tomó el aviso
      description: Idempotente. Solo con la alerta activa.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Estado de la alerta con `central_911_confirmada:true`
          content: { application/json: { schema: { $ref: '#/components/schemas/EstadoAlerta' } } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }

  # ---------------------------------------------------------------- Salud (fuera de /api/v1)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 'Token de 64 hex que devuelven `/auth/login` y `/auth/clave`. Vale 24 h (RNF-03).'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    Sesion:
      description: Sesión de 24 h
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Sesion' }
    DatosInvalidos:
      description: Falta un campo o tiene formato incorrecto.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: datos_invalidos, mensaje: 'presente debe ser true o false.' } }
    NoAutorizado:
      description: Token ausente, inválido o vencido (o credenciales incorrectas en /auth/login).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_autorizado, mensaje: 'Sesión vencida. Volvé a iniciar sesión.' } }
    Prohibido:
      description: El vecino no tiene permiso (p. ej. no es del comité).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: prohibido, mensaje: 'Solo el comité puede hacer esto.' } }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_encontrado, mensaje: 'La alerta no existe.' } }
    Conflicto:
      description: El estado actual no permite la operación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: conflicto, mensaje: 'La alerta ya no está activa.' } }
    DemasiadosIntentos:
      description: Más de `AUTH_RATE_LIMIT_MAX` (20) pedidos a /auth/* desde la misma IP en 15 minutos.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: demasiados_intentos, mensaje: 'Demasiados intentos. Probá de nuevo en unos minutos.' } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum: [datos_invalidos, no_autorizado, prohibido, no_encontrado, conflicto, demasiados_intentos, error_interno]
            mensaje: { type: string, description: 'Texto en castellano, apto para mostrar al usuario.' }
    Rol:
      type: string
      enum: [vecino, vigia, comite]
      default: vecino
    Sesion:
      type: object
      required: [token, expira_en, vecino]
      properties:
        token: { type: string, pattern: '^[0-9a-f]{64}$' }
        expira_en: { type: string, format: date-time, example: '2026-09-24T19:00:00Z' }
        vecino:
          type: object
          required: [id, nombre, manzana, rol]
          properties:
            id: { type: integer, example: 1 }
            nombre: { type: string, example: María Ríos, description: Nombre y apellido }
            manzana: { type: string, example: Manzana 4 }
            rol: { $ref: '#/components/schemas/Rol' }
...
    AlertaCerrada:
      type: object
      required: [id, estado]
      properties:
        id: { type: integer }
        estado: { type: string, enum: [activa, cancelada, resuelta] }
    EstadoAlerta:
      type: object
      description: |
        Los cinco primeros campos son los que ya lee `EstadoAlerta.fromJson` (lib/models/alerta.dart);
        el resto es nuevo en v1.1.
      required: [id, estado, central_911_confirmada, vecinos_avisados, en_camino, mi_respuesta, lat, lng, creada_en, cerrada_en, vecino]
      properties:
        id: { type: integer, example: 17 }
        estado: { type: string, enum: [activa, cancelada, resuelta] }
        central_911_confirmada: { type: boolean }
        vecinos_avisados: { type: integer, example: 38 }
        en_camino: { type: integer, example: 3 }
        mi_respuesta:
          type: [string, 'null']
          enum: [notificado, en_camino, null]
          description: null si quien consulta no fue avisado (p. ej. es quien la disparó)
        lat: { type: [number, 'null'] }
        lng: { type: [number, 'null'] }
        creada_en: { type: string, format: date-time }
        cerrada_en: { type: [string, 'null'], format: date-time }
        vecino:
          type: object
          description: Quien disparó la alerta
          required: [id, nombre, manzana]
          properties:
            id: { type: integer }
            nombre: { type: string, example: María Ríos }
            manzana: { type: string, example: Manzana 4 }